api/requirements.txt
flask
flask-cors
flask-restplus
flask-swagger-ui
Flask-SQLAlchemy 
Flask-Login 
PyJWT
google-cloud-documentai
pandas